refactored profiling `prof` into executable package
