+++ prof/__main__.py
@@ -1,0 +1,5 @@
+from .profile import main
+
+
+if __name__ == '__main__':
+    main(prog='python -m prof')
